solver: always cache the original reference instance
